hey color ekstrakto works now

+++ server/src/app.js
@@ -73,12 +73,13 @@ app.get('/colors', (req, res) => {
   var color_query = [];
   const exec = util.promisify(require('child_process').exec);
   (async () => {
-      const {stdout, stderr} = await exec('ek --number-of-colors 5 image.png');
+      const {stdout, stderr} = await exec('ek image.png');
       color_query = JSON.parse(stdout);
       console.log(color_query);
       res.send({ color_query });
 
-      if (color_query === undefined || color_query.length == 0) {  
+      // Checking if an image has been processed
+      if (color_query !== undefined && color_query.length != 0) {  
         // Clear Image after processing
         fs.unlink('image.png', (err) => {
           if (err) throw err;
